refactor(blorktools): log report move
